freelancescrapper/requirements.txt
python-dotenv==1.0.0
scrapy==2.11.0
